refactor: rename tax advantaged settings folder
