+++ frontend/src/settings/components/TaxAdvantaged/TaxAdvantagedCategoriesSection/OpeningUsageLabel.tsx
